revert GsplatRenderer and add GsplatGroupRenderer
