+++ Runtime/GsplatGroupRenderer.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ce0eaa90dbf24670a230447dc3ec09f2
+guid: eb2e3f309823f3de2a232f7046aece53
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
