+++ Runtime/GsplatRendererImpl.cs
@@ -115,8 +115,29 @@ namespace Gsplat
         /// <param name="layer">Layer used for rendering.</param>
         /// <param name="gammaToLinear">Covert color space from Gamma to Linear.</param>
         /// <param name="shDegree">Order of SH coefficients used for rendering. The final value is capped by the SHBands property.</param>
-        public void Render(uint splatCount, Transform transform, Bounds worldBounds, int layer,
+        public void Render(uint splatCount, Transform transform, Bounds localBounds, int layer,
             bool gammaToLinear = false, int shDegree = 3)
+        {
+            if (!Valid || !GsplatSettings.Instance.Valid || !GsplatSorter.Instance.Valid)
+                return;
+
+            m_propertyBlock.SetInteger(k_splatCount, (int)splatCount);
+            m_propertyBlock.SetInteger(k_gammaToLinear, gammaToLinear ? 1 : 0);
+            m_propertyBlock.SetInteger(k_splatInstanceSize, (int)GsplatSettings.Instance.SplatInstanceSize);
+            m_propertyBlock.SetInteger(k_shDegree, shDegree);
+            m_propertyBlock.SetMatrix(k_matrixM, transform.localToWorldMatrix);
+            var rp = new RenderParams(GsplatSettings.Instance.Materials[SHBands])
+            {
+                worldBounds = GsplatUtils.CalcWorldBounds(localBounds, transform),
+                matProps = m_propertyBlock,
+                layer = layer
+            };
+
+            Graphics.RenderMeshPrimitives(rp, GsplatSettings.Instance.Mesh, 0,
+                Mathf.CeilToInt(splatCount / (float)GsplatSettings.Instance.SplatInstanceSize));
+        }
+
+        public void Render(uint splatCount, Bounds worldBounds, int layer, bool gammaToLinear = false, int shDegree = 3)
         {
             if (!Valid || !GsplatSettings.Instance.Valid || !GsplatSorter.Instance.Valid)
                 return;
