+++ Runtime/GsplatUtils.cs
@@ -53,12 +53,12 @@ namespace Gsplat
 
         public static Quaternion ToQuaternion(this Vector4 v)
         {
-            return new Quaternion(v.x, v.y, v.z, v.w);
+            return new Quaternion(v.y, v.z, v.w, v.x);
         }
 
         public static Vector4 ToVector4(this Quaternion q)
         {
-            return new Vector4(q.x, q.y, q.z, q.w);
+            return new Vector4(q.w, q.x, q.y, q.z);
         }
     }
 }
